% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Set Functions}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.}

\item{table}{vector or \code{NULL}: the values to be matched against.}

\item{nomatch}{the values to be returned in the case when no match is found. Note that it is coerced to integer.}
}
\description{
A collection of pipe-able set functions.
}
