% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkour.R
\name{parkour}
\alias{parkour}
\title{Parkour}
\usage{
parkour(parmethod = "localpar")
}
\arguments{
\item{parmethod}{vector or \code{NULL}: the values to be matched.}
}
\description{
A function to automatically create a local parallel environment.
}
