% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilt-hla.R
\name{QUILT_HLA}
\alias{QUILT_HLA}
\title{QUILT_HLA}
\usage{
QUILT_HLA(
  bamfile,
  region,
  outputdir = "",
  prepared_hla_reference_dir = "",
  quilt_hla_haplotype_panelfile = "",
  finaloutputfile = NA,
  overrideoutput = FALSE,
  nGibbsSamples = 15,
  n_seek_iterations = 3,
  quilt_seed = NA,
  chr = "chr6",
  quilt_buffer = 5e+05,
  quilt_bqFilter = 10
)
}
\arguments{
\item{bamfile}{Path to bamfile to analyze}

\item{region}{HLA region to be analyzed, for example A for HLA-A}

\item{outputdir}{Optional, what output directory to use. Use only if not specifying finaloutputfile}

\item{prepared_hla_reference_dir}{Output directory containing HLA reference material necessary for QUILT HLA}

\item{quilt_hla_haplotype_panelfile}{Prepared HLA haplotype reference panel file}

\item{finaloutputfile}{Final output file path}

\item{nGibbsSamples}{How many QUILT Gibbs samples to perform}

\item{n_seek_iterations}{How many seek iterations to use in QUILT Gibbs sampling}

\item{quilt_seed}{When running QUILT Gibbs sampling, what seed to use, optionally}

\item{chr}{What chromosome, probably chr6 or maybe 6}

\item{quilt_buffer}{For QUILT Gibbs sampling, what buffer to include around center of gene}

\item{quilt_bqFilter}{For QUILT Gibbs sampling, do not consider sequence information if the base quality is below this threshold}
}
\value{
Results in properly formatted version
}
\description{
QUILT_HLA
}
\author{
Robert Davies
}
