% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilt-hla-prepare-reference.R
\name{QUILT_HLA_prepare_reference}
\alias{QUILT_HLA_prepare_reference}
\title{QUILT_HLA_prepare_reference}
\usage{
QUILT_HLA_prepare_reference(
  outputdir,
  ipd_igmt_alignments_zip_file,
  quilt_hla_supplementary_info_file,
  all_hla_regions = c("A", "B", "C", "DMA", "DMB", "DOA", "DOB", "DPA1", "DPA2", "DPB1",
    "DPB2", "DQA1", "DQA2", "DQB1", "DRA", "DRB1", "DRB3", "DRB4", "DRB5", "E", "F", "G",
    "HFE", "H", "J", "K", "L", "MICA", "MICB", "N", "P", "S", "TAP1", "TAP2", "T", "U",
    "V", "W", "Y"),
  hla_regions_to_prepare = c("A", "B", "C", "DQA1", "DQB1", "DRB1")
)
}
\arguments{
\item{outputdir}{What output directory to use}

\item{ipd_igmt_alignments_zip_file}{Path to zip file with alignments from IPD-IGMT (see README and example for more details)}

\item{quilt_hla_supplementary_info_file}{Path to file with supplementary information about the genes, necessary for proper converstion. File is tab separated without header, with 4 columns, with the following entries. First is the HLA gene name (e.g. A for HLA-A). Second is the correponding row in the _gen.txt IPD-IMGT file. Third is the position of the first column in the reference genome. Fourth is the strand (options 1 or -1).}

\item{all_hla_regions}{Character vector of all HLA genes for which IPD-IMGT files are available for}

\item{hla_regions_to_prepare}{Character vector of HLA genes to prepare for imputation}
}
\value{
Results in properly formatted version
}
\description{
QUILT_HLA_prepare_reference
}
\author{
Robert Davies
}
