% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilt-prepare-reference.R
\name{QUILT_prepare_reference}
\alias{QUILT_prepare_reference}
\title{QUILT_prepare_reference}
\usage{
QUILT_prepare_reference(
  outputdir,
  chr,
  nGen,
  regionStart = NA,
  regionEnd = NA,
  buffer = NA,
  output_file = "",
  reference_vcf_file = "",
  reference_haplotype_file = "",
  reference_legend_file = "",
  reference_sample_file = "",
  reference_populations = NA,
  reference_phred = 30,
  reference_exclude_samplelist_file = "",
  region_exclude_file = "",
  genetic_map_file = "",
  nMaxDH = NA,
  tempdir = NA,
  make_fake_vcf_with_sites_list = FALSE,
  output_sites_filename = NA,
  expRate = 1,
  maxRate = 100,
  minRate = 0.1,
  use_mspbwt = FALSE,
  mspbwt_nindices = 4L,
  override_use_eMatDH_special_symbols = NA,
  use_hapMatcherR = TRUE,
  impute_rare_common = FALSE,
  rare_af_threshold = 0.001,
  use_list_of_columns_of_A = TRUE
)
}
\arguments{
\item{outputdir}{What output directory to use}

\item{chr}{What chromosome to run. Should match BAM headers}

\item{nGen}{Number of generations since founding or mixing. Note that the algorithm is relatively robust to this. Use nGen = 4 * Ne / K if unsure}

\item{regionStart}{When running imputation, where to start from. The 1-based position x is kept if regionStart <= x <= regionEnd}

\item{regionEnd}{When running imputation, where to stop}

\item{buffer}{Buffer of region to perform imputation over. So imputation is run form regionStart-buffer to regionEnd+buffer, and reported for regionStart to regionEnd, including the bases of regionStart and regionEnd}

\item{output_file}{Path to output RData file containing prepared haplotypes (has default value that works with QUILT)}

\item{reference_vcf_file}{Path to reference haplotype file in VCF format (values must be 0 or 1)}

\item{reference_haplotype_file}{Path to reference haplotype file in IMPUTE format (file with no header and no rownames, one row per SNP, one column per reference haplotype, space separated, values must be 0 or 1)}

\item{reference_legend_file}{Path to reference haplotype legend file in IMPUTE format (file with one row per SNP, and a header including position for the physical position in 1 based coordinates, a0 for the reference allele, and a1 for the alternate allele)}

\item{reference_sample_file}{Path to reference sample file (file with header, one must be POP, corresponding to populations that can be specified using reference_populations)}

\item{reference_populations}{Vector with character populations to include from reference_sample_file e.g. CHB, CHS}

\item{reference_phred}{Phred scaled likelihood or an error of reference haplotype. Higher means more confidence in reference haplotype genotypes, lower means less confidence}

\item{reference_exclude_samplelist_file}{File with one column of samples to exclude from reference samples e.g. in validation, the samples you are imputing}

\item{region_exclude_file}{File with regions to exclude from constructing the reference panel. Particularly useful for QUILT_HLA, where you want to exclude SNPs in the HLA genes themselves, so that reads contribute either to the read mapping or state inference. This file is space separated with a header of Name, Chr, Start and End, with Name being the HLA gene name (e.g. HLA-A), Chr being the chromosome (e.g. chr6), and Start and End are the 1-based starts and ends of the genes (i.e. where we don't want to consider SNPs for the Gibbs sampling state inference)}

\item{genetic_map_file}{Path to file with genetic map information, a file with 3 white-space delimited entries giving position (1-based), genetic rate map in cM/Mbp, and genetic map in cM. If no file included, rate is based on physical distance and expected rate (expRate)}

\item{nMaxDH}{Integer Maximum number of distinct haplotypes to store in reduced form. Recommended to keep as 2 ** N - 1 where N is an integer greater than 0 i.e. 255, 511, etc}

\item{tempdir}{What directory to use as temporary directory. If set to NA, use default R tempdir. If possible, use ramdisk, like /dev/shm/}

\item{make_fake_vcf_with_sites_list}{Whether to output a list of sites as a minimal VCF, for example to use with GATK 3 to genotype given sites}

\item{output_sites_filename}{If make_fake_vcf_with_sites_list is TRUE, optional desired filename where to output sites VCF}

\item{expRate}{Expected recombination rate in cM/Mb}

\item{maxRate}{Maximum recomb rate cM/Mb}

\item{minRate}{Minimum recomb rate cM/Mb}

\item{use_mspbwt}{Build mspbwt indices to be used in imputation}

\item{mspbwt_nindices}{How many mspbwt indices to build}

\item{override_use_eMatDH_special_symbols}{Not for general use. If NA will choose version appropriately depending on whether a PBWT flavour is used.}

\item{use_hapMatcherR}{Used for nMaxDH less than or equal to 255. Use R raw format to hold hapMatcherR. Lowers RAM use}

\item{impute_rare_common}{Whether to use common SNPs first for imputation, followed by a round of rare imputation}

\item{rare_af_threshold}{Allele frequency threshold under which SNPs are considered rare, otherwise they are considered common.}

\item{use_list_of_columns_of_A}{If when using mspbwt, use columns of A rather than the whole thing, to speed up this version}
}
\value{
Results in properly formatted version
}
\description{
QUILT_prepare_reference
}
\author{
Robert Davies
}
