% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipefuns.R
\name{rename_columns}
\alias{rename_columns}
\title{Pipe Functions}
\usage{
rename_columns(data, names)
}
\arguments{
\item{data}{Input data with columns or rows to be renamed.}

\item{names}{Vector of characters to}
}
\description{
A set of transformations altered for use in the pipe.
}
