% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{transpose}
\alias{transpose}
\title{Transpose}
\usage{
transpose(data, rownames = FALSE)
}
\arguments{
\item{data}{matrix, dataframe, tibble, or list to be transposed.}

\item{rownames}{Should the output data use the rownames from the input data as columns?
Logical (Default = \code{FALSE})}
}
\description{
A type-specific version of the traditional transpose `t()` function.
}
