% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilt.R
\name{QUILT}
\alias{QUILT}
\title{QUILT}
\usage{
QUILT(chr)
}
\arguments{
\item{chr}{What chromosome to run. Should match BAM headers}
}
\value{
Results in properly formatted version
}
\description{
QUILT
}
\author{
Robert Davies
}
