% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilt-hla.R
\name{QUILT_HLA}
\alias{QUILT_HLA}
\title{QUILT_HLA}
\usage{
QUILT_HLA(
  bamlist,
  region,
  dict_file,
  hla_gene_region_file = NULL,
  outputdir = "",
  summary_output_file_prefix = "quilt.hla.output",
  nCores = 1,
  prepared_hla_reference_dir = "",
  quilt_hla_haplotype_panelfile = "",
  final_output_RData_file = NA,
  write_summary_text_files = TRUE,
  overrideoutput = FALSE,
  nGibbsSamples = 15,
  n_seek_iterations = 3,
  quilt_seed = NA,
  chr = "chr6",
  quilt_buffer = 5e+05,
  quilt_bqFilter = 10,
  summary_best_alleles_threshold = 0.99,
  downsampleToCov = 30
)
}
\arguments{
\item{bamlist}{Path to file with bam file locations. File is one row per entry, path to bam files. Bam index files should exist in same directory as for each bam, suffixed either .bam.bai or .bai}

\item{region}{HLA region to be analyzed, for example A for HLA-A}

\item{dict_file}{Path to dictionary file for reference build}

\item{hla_gene_region_file}{For reference packages built after QUILT 1.0.2, this is not used. For older reference packages, this is needed, and is a path to file with gene boundaries. 4 columns, named Name Chr Start End, with respectively gene name (e.g. HLA-A), chromsome (e.g. chr6), and 1 based start and end positions of gene}

\item{outputdir}{What output directory to use. Otherwise defaults to current directory}

\item{summary_output_file_prefix}{Prefix for output text summary files}

\item{nCores}{How many cores to use}

\item{prepared_hla_reference_dir}{Output directory containing HLA reference material necessary for QUILT HLA}

\item{quilt_hla_haplotype_panelfile}{Prepared HLA haplotype reference panel file}

\item{final_output_RData_file}{Final output RData file path, if desired}

\item{write_summary_text_files}{Whether to write out final summary text files or not}

\item{nGibbsSamples}{How many QUILT Gibbs samples to perform}

\item{n_seek_iterations}{How many seek iterations to use in QUILT Gibbs sampling}

\item{quilt_seed}{When running QUILT Gibbs sampling, what seed to use, optionally}

\item{chr}{What chromosome, probably chr6 or maybe 6}

\item{quilt_buffer}{For QUILT Gibbs sampling, what buffer to include around center of gene}

\item{quilt_bqFilter}{For QUILT Gibbs sampling, do not consider sequence information if the base quality is below this threshold}

\item{summary_best_alleles_threshold}{When reporting results, give results until posterior probability exceeds this value}

\item{downsampleToCov}{For imputing states specifically using QUILT, what coverage to downsample individual sites to. This ensures no floating point errors at sites with really high coverage. This is not used in the direct read mapping}
}
\value{
Results in properly formatted version
}
\description{
QUILT_HLA
}
\author{
Robert Davies
}
